agents/discovery_agent/requirements.txt
# Core dependencies
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
kafka-python>=2.0.2
prometheus-client>=0.16.0
pydantic>=2.0.0
PyYAML>=6.0
redis>=4.5.0
tenacity>=8.2.0

# DSPy and LLM dependencies
dspy-ai>=2.4.0
google-ai-generativelanguage>=0.4.0
google-api-core>=2.11.0
google-auth>=2.17.0
google-generativeai>=0.3.0

# Additional utilities
lxml>=4.9.0
requests>=2.28.0
urllib3>=1.26.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0